Imported TextMeshPro, added GeneEditingMenu scene with basic UI
I added a scene which displays basic UI elements for the baby character creation menu. This is called a GeneEditingMenu for now, as the speculative technology is supposed to be about gene editing and parenting choices in the future. Well, we have a nice title, a few checkboxes for the baby's sex, and two sliders for desired adult height and skin color. I also wanted to give a bit of personality to the menu so I added a placeholder name for the fake software: DreamBabySoftware (version 1.3.1). Doing all of this character creation stuff inspired from MMOs and 3D modelling reminded me to add a little baby preview screen to the right, which would help the player to make these life-altering choices obviously.