Binded the togglers for the baby's sex and the sliders for the baby's skin color to placeholder baby 3D model
Changing the sex of the baby through a single click feels like power already. It's only printing it via the console log, but it's binded. Same goes for the baby's adult height and skin color. The skin color sliders in particular are interesting - the continuum of float values that the baby can have seems to do a good job at objectifying (reducing) the baby to a material. I think there may be potential in the act of changing a baby skin's color via sliders, but the whole edit doesn't feel satisfying yet. However, providing continuous values for the skin seems a better deal than selecting individually different color boxes as one usually does in many games, as there is more freedom. I didn't think making the same deal for the baby's sex was a good idea though, that would add to much complexity but it would be entirely possible (e.g., anima/animus). The block placeholder that shows the baby is obviously underwhelming to look at, and I'm wondering if it would be a nice surprise for the player to only see the baby's final look after all the choices are made versus at each step like this.