Added controls for the gene editing menu, set canvas to world renderer and changed the 3D orientation of the canvas (slight horizontal tilt futuristic effect). Added baby preview
This edit changes the visual appearance of the menu completely. The canvas is now set in world orientation instead of overlay which makes it possible to incline it and give a futuristic or modern effect. The problem of the baby 3D model preview was reframed as a consequence of this - it previously was about embedding a miniature camera scene inside the canvas overlay, but now that the canvas is set to world, the baby's model just simply shares the 3D space with the canvas, like happy friends. The modern feel of this menu makes the presentation less direct and serious (slight playfulness due to the canvas tilt? size of baby model?)? Don't know yet if it's a good thing. Also bundling the baby's customizations per general category (e.g., biological data, socioeconomic status) seemed a better idea than having the player have to go through the whole sequence of pages - if there were 5 pages and they needed to get to page 3 they would have to click next or previous 2 times from page 1 or 5 at worst, and they'd probably forget what each page was about.