Added sound effects for newborn's gender toggles, and narrators reading their respective wikipedia entry; Refactored codebase into MVC architecture; Added support for newborn head and torso mesh changes;  Added placeholder psychological and socioeconomic edit pages
I started adding sounds and narration prematurely to investigate whether the experience of sampling through a newborn's gender options could already cause a different qualia or sensation (See repos' conversations wiki page for more details) before committing to add many more customization choices. The narrator reads the wiki's first paragraph description of each option, using Microsoft's Azure neural AI-based text-to-speech that I recorded using Audacity. The desired effect is to slow down the selection process through the edits to pay attention to what is being said, and to be more deliberate about each choice. I may have to add the text to the side in the future if the effect doesn't quite work or the content of the speech is not as comprehensible as intended. To test later. Another change is that I put discrete options for the newborn's head and torso mesh. Is this necessary? That's also unknown, although the intention was to make the player be more attached/involved with their newborn. Although highly strange, it also allows the possibility of adding a slider for genital size to drive home the point of the strangeness of gene editing a new born's biology and could ostentially be humorous. The last design specific remark is on the two new edit tabs: psychological and socioeconomic. These were inevitable as they represent two other big areas of influence in a person's life, although I'm not sure if I should change the terms belonging to real-life into more game-design appropriate, accessible terms. For example, instead of "temperament" or "extraversion/introversion" traits, the game version might be "charisma" or more RPG-like terms such as "leader" or "loud vs. silent". For occupations in the socioeconomic tab, it's unsure if it should be kept and whether the terms need to match real-life occupations as well - this is a theme issue. For now, the theme is a sci-fi futuristic Mars colonization theme (or possibly another galaxy, etc.), so occupations may need to match that speculative era. Right now they start from the stone age?