Added scriptable object for 3D assets; Adde base async JSON save system; Added base tooltip system; Added base identification tab for the newborn, and initial psychological parameters; Added stats for the colonist;  Changed the design of the view (monochrome)
This commit is a transition between version 1 and version 2 of the game, while adding a few small game systems mentioned in the title. The main one being the save system, which allows for data persistence. The shift comes from realizing that version 1 seems a bit contrived? I felt the need to add a narrative framework around the game to make it more of a "complete" game experience. This is entirely subjective, but also related to my interests in creating a more accessible, marketable game of its own. I'll have to watch out for the scope and feature creep, but for now, it seems the fake software will have a few more pages outside of the newborn edit, and work around the theme of death and parenting in a more hollistic way. At least that's the intention.