Added base psychological traits editing choices; Added FPS starter; Added greybox placeholder space station 3D environment; Moved the character creation UI in there; Put starfield skybox
I added base psychological traits for editing belonging to trait theory the big five. Basically a really supported view of lifespan and personality development. Not sure how it's going to play out - pretty complex to deal with, but I'm curious to see how it might be used in game design? So far I haven't played a game with explicit editable psychological variables, except DnD traits. Aside, I've added the WIP 3D space station and colonist pods that I'm modelling in Blender, along with a starfield skybox to make it like it's in space. I also imported the starter FPS from Unity to be able to move around. The design issue is whether the Presence and immersive factor of this extra layer is worth it. I appreciate the sci-fi-ness of seeing the character creation UI float in air like a holograph though. The next issue will be to make it raycastable/interactable like this.