Added world-space interaction with creation menu (overlay vs. world); Added player states; Added player reset location scriptable object
This commit brings together the newborn edit menu of version 1 and the placeholder 3D virtual world of version 2. Speculation is that it generates new conceptual clicks from making the player play as a double-entity : Creator and Creature. This would enable interesting play around metaphysical aspects of gameplay, to be generalized for any creation-endeavor that happens in-game, and ultimately create an artistic metaphor for Man vs. God components outside the game, which is the intended desire.