Added clock system and ship event logger, and lethal events to test death.
I added a clock system and event logger in order to test out lethal events and what to do with the dead colonists. This is primarily adding another set of publisher/subscriber pairs between the clock controller, the colonist's (HP) and the dashboard OS where the event logger resides. The event logger has a fixed (big height of 5000 units) size for now but there is a queue to process events - once the cap is reached (100 for now), the first logs are dequeued to create some space. I mainly have problems with instantiating new text components in the scroll rect at runtime (there's a lot of things that can go wrong), so for now all the logs are concatened in a single text component.