Put the character creator inside the dashboard UI; Refined the UI elements' layouts; Placed the one liner poem thing inside the character creator; Added Bot AI; Added start of character camera feed tracks; Removed character movement
I put the character creator in the dashboard's home page (called bridge), next to the event log to make it easy to contrast the life vs. death idea. It's also less awkward and disconnected. The placement of camera tracks inside the UI using render textures also fits the idea of watching life and watching things die. Instead of going to the UI in the world, and walking it, the UI tracks the player's creations in the world. Seemed like a good idea.