Added tombstone watching behaviour AI to bots; Death sculpture spawns at death location;  Fix to chat pending callers; Start of shake cam on death; Added main menu scene; Added notification controller; Deaths trigger notification; Changed language in each page, replaced and resized UI elements; Reduced max participants to 3; Added start of item system
This commit adds more deliberate, sober intentionality to death events and how they are enacted through the created bots. The bots have some rudimentary intelligence, but it's about reacting to death and others' death, and how these momentanous feedbacks operate on the player experience. Not 100% about the language yet, but I like putting the word "observe" on the On Air page, for some reason. I was afraid that it'd become too "video-gamy" and detract from the intention of using Camera as Art. The notifications is also an effect to create feedback at the right timing: on death. The next step is to add viewer stream reactions to it, too. It's about reactions. I also added some code for using items to interact with the participants through the camera feeds - dropping aids or traps like in Hunger Games seemed to be a promising and entertaining idea. A big issue is that now it's kind of decided that the player doesn't actively create new participants, but he also reacts semi-passively to the process (lifestream) - in line with the passivity of watching streams but also being an interactor - and creates when called upon. The keyword is "Streaming", like a water stream. That is the new artistic fixation. "Reacting and Streaming". So, dropping items on participants to help or hinder them is the only "active-ing thing", but also I need to give something to potentially do in between new inputstreams (characters coming in) - or maybe it's already enough to watch.