Added test "Snake" Injury Quadrant Event; Added support for "Move To Quadrant X Events"
Geez, that was complicated.

I had to the buttons with an event that mapped the correct character to their neighbouring (valid) camera feed quadrants.

I added this to see what the experience of cueing characters about potential opportunities happening around them (or traps in this case, with the Snake quadrant event) would feel like.