Added real game terrain; added base male (headless) with idle and walking animation; Added weather system; Adjusted audition page and various fixes; Setup camera lanes in quadrant at various angles
This commit packs in the art big time, and it's bringing much needed life. I should have dropped these art assets sooner. It really does change the perceived feelings (and that could even re-adjust priorities?). Anyways, we have a bunch of headless people walking around in a foggy, mystical landscape of four seasons. It somehow feels spiritual to watch headless beings walk in a barren landscape. This commit bears a few problems with the nav agent setup on the navmesh so sometimes the character bots get stuck in place (will fix that next).