Added viewers and subscribers; Added special events feed; Added REACTIONS FX, notification, stream chat bot, top 10 donator money ranking; Bot navigation is more stable; Bots spawn on special lift at start
The chatbots in the stream live chat are a bit offensive and vulgar, but it's pretty representative of the real world so I left it there. I would probably release a PG safe version of the game eventually (some of the texts from the corpus are very sensitive). Also it's a bit funny and that's sorely needed probably. I added a special camera feed showing the intro and outro of the season's phases. I still need to put individual textures for seeing audition actors before they get cast. I also added the very satisfying FX for new subscribers and donators to the player's fake channel. I also added a top 10 donators and donation money total tracker on the bridge's page. There are new adverbs for the event log too.  Technically-wise, the bots fare a bit better with Unity's Navmesh system, but sometimes they get stuck in terrain still, so it would require more agent tweaking (max slope, etc.). I put as failsafe that the buttons originating the unsuccessful navigation do not get removed if that occurs (during quadrant selection and also ui action event). Finally, I also changed the action method pointers in the waypoint events to actual delegates with return types so I can get a bool indicating successful navigation or not. The audition editor spawning should also be stable now.  Note: I'm happy to have done a bit of functional programming for the donator ranking and it was pretty fun and satisfying.