Added subscriber requests pop-ups; Added gold and snakes at waypoints spawning;  Added album covers in Jukebox and news pages; Changed the typeface to Noto Sans JP and Yu Mincho
This commit makes the software look a bit more modern, but WITHOUT THE HOLOGRAM ShAderGrapH on the UI and characters it's not the final look yet. Not completely sure if it's better than the last font  or if I'm just fatigued... Anyways, I also added subscriber requests pop-ups and it's clear now that the experience is very much more about fake streaming than before. I wonder how to parse the in-game experience now (the death game on live things)? An idea is to focus on the performativity of the characters - add dance animations (on the way already), death tribute animations, singing somehow, more on the performative and camera art angle than the "game mechanics design" angle than previously focused.