Slayer behaviour update (seeks and stops at predators); action buttons chime sounds test; new geometric action icons; Revert to terrain collider
The slayer seeks and stops to do something to predators (presumably attack it next). Not sure if going to make it funnier and have the snake run away instead, but this was an improvement on the AI. I also added chime sounds on the action icons that were also changed to geometric shapes (aka, ps4) as a test. It just looks good when the icon is dragged and scaled up + made transparent, but it looked silly with a snake or shovel icon. Never mind the bread icon. Not sure what to think about representation of actions like this - but I like the idea of putting a controller manual in the walkthrough for the buttons. Also I removed the prefab terrain that was made from Blender as the collider on it along with its generated navmesh generally sucked. I'll consider that gray boxing anyways. The built-in terrain with its collider in Unity performs much better on all aspects (navigation, collision, speed), so that had to be done since I couldn't get some AI navigation right. The last annoying thing is that for now the actors are spawned with rigidbody and gravity and they can get tripped up with other bodies like the actors - I hiked up the mass factor and it's not a big issue anymore but it's something to take note of. Also the physics is more satisfying than the kinematic version.